+++ src/api/activity/interface.ts
@@ -1,9 +1,12 @@
+import { IPageInfo } from '../../common/Inteface';
+
 export interface FetchActivityListReq {
   size: number;
 }
 
 export interface FetchActivityListRes {
-  dataList: any[]
+  pageInfo:IPageInfo
+  dataList: IActivityItem[]
 }
 export interface IActivityItem {
   img: string;
@@ -11,12 +14,12 @@ export interface IActivityItem {
   summary: string;
   pubTime: string;
   explore: number;
-  id:number;
+  id: number;
 }
-export interface IActivityDetail{
-  title:string;
-  img:string;
-  content:string;
-  pubTime:string;
-  explore:number;
+export interface IActivityDetail {
+  title: string;
+  img: string;
+  content: string;
+  pubTime: string;
+  explore: number;
 }
