+++ src/common/Inteface.ts
@@ -1,0 +1,9 @@
+export interface IPageInfo {
+  everyPage: number;
+  totalCount: number;
+  currentpage: number;
+  beginIndex: number;
+  hasPrePage: boolean;
+  hasNextPage: boolean;
+  totalPage: number;
+}
