+++ src/pages/activity/index.less
@@ -1,34 +1,35 @@
-.wiki{
+.activity{
   padding-top: 70px;
-  .wiki-header{
+  .activity-header{
     background-image: url('http://dummyimage.com/200x100');
     background-size: cover;
     background-repeat: no-repeat;
     height: 320px;
-    .wiki-title{
+    .activity-title{
       position: relative;
       top: 100px;
       left: 160px;
       font-size: 36px;
       font-style: oblique;
     }
-    .wiki-describe{
+    .activity-describe{
       position: relative;
       top: 90px;
       left: 160px;
       font-size: 18px;
     }
   }
-  .wiki-body{
+  .activity-body{
+    margin-top: 50px;
     display: flex;
     flex-direction: column;
     justify-content: center;
     align-items: center;
   }
-  .wiki-bottom{
+  .activity-bottom{
     display: flex;
     justify-content: center;
-    margin-top: -50px;
-    margin-bottom: 100px;
+    margin-top: 50px;
+    margin-bottom: 50px;
   }
 }
